+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Public profile page which can be shared with friends
   - Public profile link has a nice messenger preview
   - Accept/decline/withdraw friend requests
+  - Friend profile statistics screen
   - Deep linking: profile links open the Glass Trail app when installed, otherwise they show the web version
   - Feed: See drink entries from friends
 - Notifications
@@ -22,9 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - friend removed
 - Track alcohol-free beer separately in drink metadata and statistics.
 - Added SonarQube analysis
-
-### Fixed
-- Delete friend drink notifications on unfriend
 
 ## [1.2.1] - 2026-04-13
 ### Fixed